refactor(elaborator): move pure type-shape helpers to TypeSystem
Stage A of the elaborator God Object decomposition (completing WEP
2026-05-26). Relocate eight pure type-table queries off impl<H>
Elaborator onto impl TypeSystem, where the WEP places "operations that
take only type IDs and names":

  build_declared_type_params, struct_name_for_type, newtype_base_lookup,
  get_ultimate_base_struct_name, get_base_type, auto_derive_eligible_kind,
  substitute_newtype_in_type, type_id_to_string

Each touches only self.type_table (now read directly instead of through
self.tysys), so both the body walk and reify reach them via self.tysys.
substitute_newtype_in_type drops its unnecessary &mut self (interior
mutability via the RefCell type table). Call sites updated across the
body-walk files. No behavior change.
