+++ wado-compiler/src/elaborator/coercion.rs
@@ -461,7 +461,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
         }
 
         // Get base struct name from target type
-        let base_name = self.struct_name_for_type(target_type)?;
+        let base_name = self.tysys.struct_name_for_type(target_type)?;
 
         // Check if target type implements KeyValueLiteralBuilder (or legacy KeyValueLiteral)
         let from_literal_info = self.find_key_value_literal_trait_impl(&base_name, target_type)?;
@@ -475,7 +475,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
         // where the monomorphizer expects to find the template. Fall back to
         // the receiver type's module for inherent / auto-derived impls; if
         // neither resolves, panic (no current-module fallback per #1110 (2)).
-        let builder_name_for_lookup = self
+        let builder_name_for_lookup = self.tysys
             .struct_name_for_type(builder_type)
             .unwrap_or_else(|| base_name.clone());
         let builder_type_module = match self.tysys.type_table.borrow().get(builder_type) {
@@ -522,7 +522,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
             .make_compiler_struct(crate::compiler_item::CompilerItem::String);
 
         // Get type args for monomorphization from builder type
-        let builder_base_name = self
+        let builder_base_name = self.tysys
             .struct_name_for_type(builder_type)
             .unwrap_or_else(|| base_name.clone());
         let (type_arg_names, type_arg_ids): (Vec<String>, Vec<TypeId>) = {
@@ -664,7 +664,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
             _ => return None,
         };
 
-        let base_name = self.struct_name_for_type(target_type)?;
+        let base_name = self.tysys.struct_name_for_type(target_type)?;
         let (seq_info, needs_newtype_cast) = self
             .find_sequence_literal_trait_impl(&base_name, target_type)
             .map(|info| (info, false))
@@ -675,7 +675,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
                     .type_table
                     .borrow()
                     .get_newtype_base(target_type)?;
-                let base_name = self.struct_name_for_type(base_type)?;
+                let base_name = self.tysys.struct_name_for_type(base_type)?;
                 self.find_sequence_literal_trait_impl(&base_name, base_type)
                     .map(|info| (info, true))
             })?;
@@ -685,7 +685,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
         let builder_type = seq_info.builder_type;
         let output_type = seq_info.output_type;
         let impl_module_source = seq_info.impl_module_source;
-        let builder_base_name = self
+        let builder_base_name = self.tysys
             .struct_name_for_type(builder_type)
             .unwrap_or_else(|| base_name.clone());
