+++ wado-compiler/src/elaborator/operators.rs
@@ -287,7 +287,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
                     match tt.get(ultimate) {
                         ResolvedType::Struct { .. } | ResolvedType::GenericInstance { .. } => {
                             drop(tt);
-                            self.struct_name_for_type(*base_type)
+                            self.tysys.struct_name_for_type(*base_type)
                         }
                         _ => None,
                     }
@@ -511,7 +511,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
 
                 // For newtypes, resolve base type for trait impl fallback
                 let (lookup_name, lookup_type_id) =
-                    self.newtype_base_lookup(&struct_name, left.type_id);
+                    self.tysys.newtype_base_lookup(&struct_name, left.type_id);
 
                 // Find the arithmetic trait implementation
                 let (trait_info_opt, impl_name) = self
@@ -607,7 +607,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
 
                 // For newtypes, resolve base type for trait impl fallback
                 let (lookup_name, lookup_type_id) =
-                    self.newtype_base_lookup(&struct_name, left.type_id);
+                    self.tysys.newtype_base_lookup(&struct_name, left.type_id);
 
                 // Find the shift trait implementation
                 let (trait_info_opt, impl_name) = self
@@ -927,7 +927,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
             };
             if let Some(struct_name) = struct_name {
                 let (lookup_name, lookup_type_id) =
-                    self.newtype_base_lookup(&struct_name, expr_type);
+                    self.tysys.newtype_base_lookup(&struct_name, expr_type);
                 let resolved = self
                     .resolve_trait_method_for_op(
                         &struct_name,
@@ -1141,7 +1141,7 @@ impl<H: CompilerHost> Elaborator<'_, H> {
 
                 // For newtypes, resolve the base type name for trait impl lookup
                 let (lookup_name, lookup_type_id) =
-                    self.newtype_base_lookup(&struct_name, base_type_id);
+                    self.tysys.newtype_base_lookup(&struct_name, base_type_id);
 
                 if !struct_name.is_empty() {
                     let index_type = self.resolve_expr(&index_expr.index, ctx, None);
